renaming the relevance_visualizer file
