+++ skater/core/visualizer/__init__.py
@@ -1,3 +1,3 @@
-from .relevance_visualizer import *
+from .text_relevance_visualizer import *
 
 __all__ = [build_explainer, generate_word_cloud, show_in_notebook]
